Support Linux in C# NuGet package

+++ odyssey_hub_client_ffi/src/ffi_common.rs
@@ -89,7 +89,7 @@ pub enum DeviceEventKind {
     DisconnectEvent,
     ZeroResult(bool),
     SaveZeroResult(bool),
-    ShotDelayChanged(u8),
+    ShotDelayChanged(u16),
     PacketEvent(PacketEvent),
 }
